requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
Authlib==1.6.8
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coqpit-config==0.2.4
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cutlet==0.5.0
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
ecdsa==0.19.1
einops==0.8.2
fastapi==0.129.0
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
fugashi==1.5.2
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
inflect==7.5.0
jaconv==0.5.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
ko-speech-tools==0.1.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mojimoji==0.0.13
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
packaging==26.0
pillow==12.0.0
platformdirs==4.9.2
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypinyin==0.55.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cu126
torchaudio==2.10.0+cu126
torchcodec==0.10.0
torchvision==0.25.0+cu126
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typeguard==4.5.1
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
yarl==1.22.0
